+++ src/components/bets/index.ts
@@ -8,5 +8,10 @@ export { BetsInfoModal } from "./BetsInfoModal";
 export { Last10BetsSection } from "./Last10BetsSection";
 export { OpenBetsSection } from "./OpenBetsSection";
 export { ProfitsAndStatisticsSection } from "./ProfitsAndStatisticsSection";
+export { PieChartCard } from "./PieChartCard";
+export { BetCountBySportChart } from "./BetCountBySportChart";
+export { MoneyBySportChart } from "./MoneyBySportChart";
+export { BetResultsChart } from "./BetResultsChart";
+export { SportAllocationChartSection } from "./SportAllocationChartSection";
 export { BetResults } from "./BetUtils";